Consolidate MuJoCo equality attributes and enums in equality.py
Move the equality-constraint custom-attribute registration into a single
home and collapse the two parallel enum hierarchies into one.

Previously eq_solref/eq_solimp were registered by SolverMuJoCo while the
projection fields (target_kind/target/objtype) were registered by
equality.py, and two enum sets coexisted (sim.enums.EqTarget/EqObjType
alongside the MuJoCo-side constants). Register every equality field in
_register_equality_constraint_attributes and keep one enum,
MjcEqualityTargetKind plus the mjtObj-aligned MJC_OBJ_* constants, in
equality.py. Delete the now-unused EqTarget/EqObjType from sim.enums.

Because the equality fields are always declared by ModelBuilder.__init__
during the deprecation window, the importer "attributes registered?"
sentinel is no longer needed; register_custom_attributes is idempotent,
so the importers call it unconditionally when conversion is requested.
